reconfiguration
