+++ api/views/application.py
@@ -43,46 +43,3 @@ def create_application():
             "message": 'Bad request parameters: {}'.format(data['message'])
         }
         return make_response(jsonify(response_object), 400)
-
-
-@application.route('/admin/applications', methods=['GET'])
-def get_all_applications():
-    """ Endpoint that gets all titles to be read by the default page """
-    all_applications = []
-    try:
-        for application in applications.find():
-            all_applications.append(application)
-
-        response_object = {
-            "status": True,
-            "allApplications": all_applications,
-            "message": 'All postings received.'
-        }
-        return make_response(jsonify(response_object), 200)
-
-    except Exception as e:
-        return make_response(return_exception(e), 400)
-
-
-@application.route('/admin/applications/<application_id>', methods=['GET'])
-def get_specific_application(application_id):
-    """ Endpoint that gets information of specific application based on id """
-    "is id default???"
-    try:
-        application_info = applications.find_one({"_id": ObjectId(application_id)})
-        if not application_info:
-            response_object = {
-                "status": False,
-                "message": 'Application ID not found.'
-            }
-            return make_response(jsonify(response_object), 404)
-
-        response_object = {
-            "status": True,
-            "applicationInfo": application_info,
-            "message": 'Application found.'
-        }
-        return make_response(jsonify(response_object), 200)
-
-    except Exception as e:
-        return make_response(return_exception(e), 400)
